+++ xocto/fields/postgres/ranges.py
@@ -4,7 +4,7 @@ Adapter fields for postgres range types that return xocto.Range python objects.
 
 import datetime
 import json
-from typing import Optional
+from typing import Any, Optional, Union
 
 from django.contrib.postgres import fields as pg_fields
 from django.contrib.postgres.fields import ranges as pg_ranges
@@ -25,15 +25,21 @@ class FiniteDateRangeField(pg_fields.DateRangeField):
     """
 
     def get_prep_value(
-        self, value: Optional[ranges.FiniteDateRange]
-    ) -> Optional[pg_ranges.DateRange]:
+        self, value: Any
+    ) -> Optional[Union[pg_ranges.DateRange, datetime.date]]:
         if value is None:
             return None
-        if not isinstance(value, ranges.FiniteDateRange):
-            raise TypeError(
-                "FiniteDateRangeField may only accept FiniteDateRange objects."
-            )
-        return pg_ranges.DateRange(lower=value.start, upper=value.end, bounds="[]")
+        if isinstance(value, ranges.FiniteDateRange):
+            return pg_ranges.DateRange(lower=value.start, upper=value.end, bounds="[]")
+        if (
+            isinstance(value, datetime.date)
+            # Don't allow datetime (datetime is a subclass of date!)
+            and not isinstance(value, datetime.datetime)
+        ):
+            return value
+        raise TypeError(
+            "FiniteDateRangeField may only accept FiniteDateRange or date objects."
+        )
 
     def from_db_value(
         self,
@@ -96,16 +102,18 @@ class FiniteDateTimeRangeField(pg_fields.DateTimeRangeField):
     """
 
     def get_prep_value(
-        self, value: Optional[ranges.FiniteDatetimeRange]
-    ) -> Optional[pg_ranges.DateTimeTZRange]:
+        self, value: Any
+    ) -> Optional[Union[pg_ranges.DateTimeTZRange, datetime.datetime]]:
         if value is None:
             return None
-        if not isinstance(value, ranges.FiniteDatetimeRange):
-            raise TypeError(
-                "FiniteDateTimeRangeField may only accept FiniteDatetimeRange objects."
+        if isinstance(value, ranges.FiniteDatetimeRange):
+            return pg_ranges.DateTimeTZRange(
+                lower=value.start, upper=value.end, bounds="[)"
             )
-        return pg_ranges.DateTimeTZRange(
-            lower=value.start, upper=value.end, bounds="[)"
+        if isinstance(value, datetime.datetime):
+            return value
+        raise TypeError(
+            "FiniteDateTimeRangeField may only accept FiniteDatetimeRange or datetime objects."
         )
 
     def from_db_value(
@@ -158,22 +166,18 @@ class HalfFiniteDateTimeRangeField(pg_fields.DateTimeRangeField):
     """
 
     def get_prep_value(
-        self, value: Optional[ranges.HalfFiniteDatetimeRange]
-    ) -> Optional[pg_ranges.DateTimeTZRange]:
+        self, value: Any
+    ) -> Optional[Union[pg_ranges.DateTimeTZRange, datetime.datetime]]:
         if value is None:
             return None
-        if (
-            # HalfFiniteDatetimeRange is a subscripted generic and may not be checked with
-            # isinstance directly. So, we check for it's parent class and attribute values.
-            not isinstance(value, ranges.HalfFiniteRange)  # type: ignore [redundant-expr]
-            or not isinstance(value.start, datetime.datetime)  # type: ignore [redundant-expr]
-            or (value.end is not None and not isinstance(value.end, datetime.datetime))
-        ):
-            raise TypeError(
-                "HalfFiniteDateTimeRangeField may only accept HalfFiniteDatetimeRange objects."
+        if self._is_half_finite_datetime_range(value):
+            return pg_ranges.DateTimeTZRange(
+                lower=value.start, upper=value.end, bounds="[)"
             )
-        return pg_ranges.DateTimeTZRange(
-            lower=value.start, upper=value.end, bounds="[)"
+        if isinstance(value, datetime.datetime):
+            return value
+        raise TypeError(
+            "HalfFiniteDateTimeRangeField may only accept HalfFiniteDateTimeRangeField or datetime objects."
         )
 
     def from_db_value(
@@ -218,3 +222,14 @@ class HalfFiniteDateTimeRangeField(pg_fields.DateTimeRangeField):
                 "end": base_field.value_to_string(end),
             }
         )
+
+    def _is_half_finite_datetime_range(self, value: Any) -> bool:
+        # HalfFiniteDatetimeRange is a subscripted generic and may not be checked with
+        # isinstance directly. So, we check for it's parent class and attribute values.
+        if not isinstance(value, ranges.HalfFiniteRange):
+            return False
+        if not isinstance(value.start, datetime.datetime):
+            return False
+        if value.end is not None and not isinstance(value.end, datetime.datetime):
+            return False
+        return True